#71 Moved GFF3 to Annotations.py
